requirements.txt
# NADIA HITL Production Requirements
# Generated for containerization and production deployment

# Core Dependencies
telethon==1.34.0               # Telegram client library
openai==1.12.0                 # OpenAI API client (GPT-4o-mini)
google-generativeai==0.8.3     # Google Gemini API client
redis==5.0.1                   # Redis client for caching and queues
python-dotenv==1.0.0           # Environment variable management
pydantic==2.11.7               # Data validation and settings
pydantic-settings==2.10.1      # Pydantic settings management

# Web Framework & API
fastapi==0.109.0               # FastAPI web framework
uvicorn==0.27.0                # ASGI server for FastAPI
slowapi==0.1.9                 # Rate limiting for FastAPI
httpx==0.28.1                  # Async HTTP client
aiohttp==3.9.3                 # Alternative async HTTP client

# Database
asyncpg==0.29.0                # PostgreSQL async driver
psycopg2-binary==2.9.9         # PostgreSQL sync driver (for migrations)

# AI and NLP
tiktoken==0.5.2                # OpenAI tokenizer for cost estimation

# System Monitoring
psutil==7.0.0                  # System and process utilities

# CLI and User Interface  
rich==14.0.0                   # Rich text and beautiful formatting
questionary==2.1.0             # Interactive command line prompts

# Utilities
PyYAML==6.0.1                  # YAML configuration files
Jinja2==3.1.2                  # Template engine

# Development and Testing
pytest==7.4.4                  # Testing framework
pytest-asyncio==0.23.3         # Async testing support
pytest-cov==4.0.0              # Code coverage
ruff==0.1.13                   # Fast Python linter and formatter

# Production Monitoring (Fase 3 additions)
structlog==24.1.0              # Structured logging
prometheus-client==0.19.0      # Prometheus metrics client
sentry-sdk[fastapi]==1.40.0    # Error tracking and monitoring

# Database Migrations
alembic==1.13.1                # Database migration tool

# Feature Flags (Fase 3)
flagsmith==3.3.0               # Feature flag management

# Security
cryptography==41.0.8           # Cryptographic recipes